# add perf tests