--- a/package.json
+++ b/package.json
@@ -24,6 +24,7 @@
     "babel-preset-es2015": "^6.18.0",
     "babel-preset-flow": "^6.23.0",
     "babel-register": "^6.26.0",
+    "benchmark": "^2.1.4",
     "check-dependencies": "^1.1.0",
     "css-loader": "^0.28.8",
     "eslint": "^4.14.0",
@@ -33,7 +34,7 @@
     "flow-bin": "^0.62.0",
     "jest": "^22.0.4",
     "jest-serializer-html": "^4.0.1",
-    "js-yaml": "^3.3.1",
+    "js-yaml": "^3.10.0",
     "json-stable-stringify": "^1.0.1",
     "jspngopt": "^0.2.0",
     "less": "~2.7.1",
@@ -66,6 +67,7 @@
     "start": "check-dependencies && webpack-dev-server --open --config webpack.dev.js",
     "build": "check-dependencies && rimraf build/* && webpack",
     "watch": "npm run build -- --watch",
+    "perf-test": "check-dependencies && NODE_ENV=test node test/perf-test.js",
     "prepublishOnly": "make NIS= dist"
   },
   "pre-commit": [