Here's the output I get when I run `npm run perf-test`:
```
compiling katex...

AccentsText x 2,416 ops/sec ±4.79% (80 runs sampled)
ArrayMode x 344 ops/sec ±1.35% (85 runs sampled)
GroupMacros x 7,374 ops/sec ±5.01% (82 runs sampled)
MathBb x 1,976 ops/sec ±1.10% (84 runs sampled)
SqrtRoot x 1,424 ops/sec ±0.95% (86 runs sampled)
StretchyAccent x 370 ops/sec ±1.64% (83 runs sampled)
Units x 298 ops/sec ±3.04% (78 runs sampled)
Fastest is GroupMacros
(khan27)kb-no-mbp:KaTeX kevinbarbash$ git st
On branch perf_test
```
This fixes #1053.