@kevinbarabash 2018-01-01
I picked tests which I thought were interesting in some way, e.g. use of macros, stretchy accents, environments, etc.  Each test takes a few seconds to run and I wasn't able to find a way to reduce the number of runs sampled so I limited the number of strings we test.