@kevinbarabash 2018-01-30
> I've already got feature requests, however:

I've added these as TODOs in perf-test.js and will created GH issues for them.