@kevinbarabash 2018-01-30
@davidflanagan thanks for the quick review.  I'm happy this in on `master` now.  It'll make understanding the perf impact of future changes much easier.